Make per-request nonce 8 bytes in size
This is _more_ than sufficient to make it infeasible to use brute force
to get data into other clients' timestamps. The limiting factor for the
attacker here is they have a 1-in-2^64 chance of guessing the correct
nonce per request.

The math is unlike nonces for data privacy, as here we're limited by
network bandwidth. 4 bytes would be ok too really. But 4 billion network
requests is hypothetically possible. So we add a few extra bytes to make
it totally infeasible.